requirements.txt
streamlit>=1.28.0
fyers-apiv3>=3.1.2
pandas>=2.0.0
numpy>=1.24.0
flask>=3.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
requests>=2.31.0
plotly>=5.17.0
ta>=0.11.0
